Implement base machine code

+++ src/main/kotlin/nichrosia/arcanology/type/tooltip/energy/EnergyTooltipData.kt
@@ -4,6 +4,6 @@ import net.minecraft.client.gui.tooltip.TooltipComponent
 import nichrosia.arcanology.type.tooltip.CustomTooltipData
 
 /** @author GabrielOlvH */
-data class EnergyTooltipData(val energy: Double, val maxEnergy: Double) : CustomTooltipData {
+data class EnergyTooltipData(val energy: Long, val maxEnergy: Long) : CustomTooltipData {
     override fun toComponent(): TooltipComponent = EnergyTooltipComponent(this)
 }
